+++ src/test/java/com/getindata/kafka/connect/iceberg/sink/testresources/TestConfig.java
@@ -2,6 +2,7 @@ package com.getindata.kafka.connect.iceberg.sink.testresources;
 
 import com.getindata.kafka.connect.iceberg.sink.IcebergSinkConfiguration;
 
+import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -12,7 +13,6 @@ public class TestConfig {
     public static final String S3_REGION_NAME = "us-east-1";
     public static final String TABLE_NAMESPACE = "debeziumevents";
     public static final String TABLE_PREFIX = "debeziumcdc_";
-    public static final String WRITE_FORMAT = "parquet";
     public static final String DEBEZIUM_CONNECT_IMAGE = "debezium/connect:1.9";
     public static final String POSTGRES_IMAGE = "postgres";
     public static final String MINIO_IMAGE = "minio/minio:latest";
@@ -25,7 +25,7 @@ public class TestConfig {
     }
 
     public static class Builder {
-        private Map<String, String> properties;
+        private final Map<String, String> properties;
 
         private Builder() {
             properties = new HashMap<>();
@@ -34,6 +34,15 @@ public class TestConfig {
             properties.put(IcebergSinkConfiguration.TABLE_PREFIX, TABLE_PREFIX);
             properties.put(IcebergSinkConfiguration.TABLE_AUTO_CREATE, "true");
             properties.put(IcebergSinkConfiguration.CATALOG_NAME, "iceberg");
+        }
+
+        public Builder withLocalCatalog(Path localWarehouseDir) {
+            properties.put(IcebergSinkConfiguration.CATALOG_TYPE, "hadoop");
+            properties.put("iceberg.warehouse", localWarehouseDir.toUri().toString());
+            return this;
+        }
+
+        public Builder withS3(String s3Url) {
             properties.put(IcebergSinkConfiguration.CATALOG_TYPE, "hadoop");
             properties.put("iceberg.fs.defaultFS", "s3a://" + S3_BUCKET);
             properties.put("iceberg.fs.s3a.endpoint.region", S3_REGION_NAME);
@@ -41,15 +50,17 @@ public class TestConfig {
             properties.put("iceberg.fs.s3a.access.key", S3_ACCESS_KEY);
             properties.put("iceberg.fs.s3a.secret.key", S3_SECRET_KEY);
             properties.put("iceberg.fs.s3a.path.style.access", "true");
-        }
-
-        public Builder withS3Url(String s3Url) {
             properties.put("iceberg.fs.s3a.endpoint", s3Url);
             return this;
         }
 
         public Builder withUpsert(boolean upsert) {
             properties.put(IcebergSinkConfiguration.UPSERT, Boolean.toString(upsert));
+            return this;
+        }
+
+        public Builder withCustomCatalogProperty(String key, String value) {
+            properties.put(IcebergSinkConfiguration.ICEBERG_PREFIX + key, value);
             return this;
         }